Added Controls enum to remove all vc configuration from App delegate
