+++ Demo/Demo/Controls.swift
@@ -1,0 +1,6 @@
+//  Copyright 2019, Oath Inc.
+//  Licensed under the terms of the MIT License. See LICENSE.md file in project root for terms.
+
+import PlayerControls
+
+enum Controls {}
